+++ internal/core/testdata/latestRelease.json
